think_blog_1103_完善后台部分功能
优化后台基础控制器的列表显示方法
优化后台基础模型的列表显示方法
js的列表显示没有数据显示(后台完成优化，前台比较混乱)
用户删除模块完善
相册删除模块完善

+++ Application/Admin/Model/MemberModel.class.php
@@ -77,38 +77,49 @@ class MemberModel extends CommonModel{
         $data['status']=0;
         $data['msg']='';
 
-        $this->head_pic=D('Member')->field('head_pic')->where(array('id'=>array('in',$this->id)))->select();   //获取用户头像，删除用户同时删除头像
+
+        //获取用户头像，删除用户同时删除头像
+        $head_pic=D('Member')->field('head_pic')->where(array('id'=>array('in',$this->id)))->select();
+        //获取相片数据，删除用户同时删除相片
+        $img_src=D('PhotoImg')->field('img_src')->where(array('member_id'=>array('in',$this->id)))->select();
         /*if(!empty(C('ROOT').C('UPLOAD').$this->head_pic) && file_exists(C('ROOT').C('UPLOAD').$this->head_pic)) {*/
-        if(count($this->head_pic)>0) {
+
             $this->startTrans();
             $article = D('Article');
             $articleType = D('ArticleType');
             $photo = D('Photo');
             $photoImg = D('PhotoImg');
+            $mess = D('Mess');
+            $complaint = D('Complaint');
             //删除本表数据放在最后，$result和$result2顺序不能改变，否则因外键约束而导致删除失败
             $result = $article->where(array('member_id' => array('in', $this->id)))->delete();
             $result2 = $articleType->where(array('member_id' => array('in', $this->id)))->delete();
             $result3 = $photoImg->where(array('member_id' => array('in', $this->id)))->delete();
             $result4 = $photo->where(array('member_id' => array('in', $this->id)))->delete();
-            $result5 = $this->where(array('id' => array('in', $this->id)))->delete();
-            if ($result!==false && $result2!==false && $result3!==false && $result4!==false && $result5!==false) {
+            $result5 = $mess->where(array('messer_id' => array('in', $this->id)))->delete();
+            $result6 = $mess->where(array('messed_id' => array('in', $this->id)))->delete();
+            $result7 = $this->where(array('id' => array('in', $this->id)))->delete();
+            if ($result!==false && $result2!==false && $result3!==false && $result4!==false && $result5!==false && $result6!==false && $result7!==false) {
                 $this->commit();
-                foreach($this->head_pic as $head_pic_arr){  //在空间中删除头像
+
+                foreach($head_pic as $head_pic_arr){  //在空间中删除头像
                     if(file_exists(C('ROOT').C('UPLOAD').$head_pic_arr['head_pic'])){
                         if(!stristr($head_pic_arr['head_pic'],'default')){
                             unlink(C('ROOT').C('UPLOAD').$head_pic_arr['head_pic']);
                         }
                     }
                 }
-                $data['msg'] = '删除成功';
+                foreach($img_src as $img_src_arr){
+                    if(file_exists(C('ROOT').C('UPLOAD').$img_src_arr['img_src'])){
+                        unlink(C('ROOT').C('UPLOAD').$img_src_arr['img_src']);
+                    }
+                }
                 $data['status'] = 1;
+                $data['msg'] = '删除成功';
             } else {
                 $this->rollback();
                 $data['msg'] = '删除失败';
             }
-        }else{
-            $data['msg'] = '用户信息获取失败';
-        }
 
         return $data;
     }
